Refactor: v4 separating queryRepository code #9
