+++ jpashop/src/main/java/jpabook/jpashop/repository/order/simplequery/SimpleOrderQueryDto.java
@@ -1,4 +1,4 @@
-package jpabook.jpashop.repository;
+package jpabook.jpashop.repository.order.simplequery;
 
 import jpabook.jpashop.domain.Address;
 import jpabook.jpashop.domain.Order;
